added compressed instruction info generation and the necessary fixes
